Added extensions for PubNubObjC.swift
