+++ Sources/PubNub/KMM/PubNubObjC+Publish.swift
@@ -1,0 +1,62 @@
+//
+//  PubNubObjC+Publish.swift
+//
+//  Copyright (c) PubNub Inc.
+//  All rights reserved.
+//
+//  This source code is licensed under the license found in the
+//  LICENSE file in the root directory of this source tree.
+//
+
+import Foundation
+
+// MARK: - Publish
+
+@objc
+public extension PubNubObjC {
+  func publish(
+    channel: String,
+    message: Any,
+    meta: Any?,
+    shouldStore: NSNumber?,
+    ttl: NSNumber?,
+    onSuccess: @escaping ((Timetoken) -> Void),
+    onFailure: @escaping ((Error) -> Void)
+  ) {
+    pubnub.publish(
+      channel: channel,
+      message: AnyJSON(message),
+      shouldStore: shouldStore?.boolValue,
+      storeTTL: shouldStore?.intValue,
+      meta: meta != nil ? AnyJSON(meta as Any) : nil
+    ) {
+      switch $0 {
+      case .success(let timetoken):
+        onSuccess(timetoken)
+      case .failure(let error):
+        onFailure(error)
+      }
+    }
+  }
+}
+
+// MARK: - Signal
+
+@objc
+public extension PubNubObjC {
+  func signal(
+    channel: String,
+    message: Any,
+    onSuccess: @escaping ((Timetoken) -> Void),
+    onFailure: @escaping ((Error) -> Void)
+  ) {
+    pubnub.signal(channel: channel, message: AnyJSON(message)) {
+      switch $0 {
+      case .success(let timetoken):
+        onSuccess(timetoken)
+      case .failure(let error):
+        onFailure(error)
+      }
+    }
+  }
+}
