+++ Sources/PubNub/KMM/PubNubObjC+Push.swift
@@ -1,0 +1,107 @@
+//
+//  PubNubObjC+Push.swift
+//
+//  Copyright (c) PubNub Inc.
+//  All rights reserved.
+//
+//  This source code is licensed under the license found in the
+//  LICENSE file in the root directory of this source tree.
+//
+
+import Foundation
+
+extension PubNubObjC {
+  func pushService(from rawString: String) -> PubNub.PushService? {
+    switch rawString {
+    case "gcm":
+      return .fcm
+    case "apns":
+      return .apns
+    case "mpns":
+      return .mpns
+    default:
+      return nil
+    }
+  }
+}
+
+@objc
+public extension PubNubObjC {
+  func addChannelsToPushNotifications(
+    channels: [String],
+    deviceId: Data,
+    pushType: String,
+    onSuccess: @escaping (([String]) -> Void),
+    onFailure: @escaping ((Error) -> Void)
+  ) {
+    guard let pushService = pushService(from: pushType) else {
+      onFailure(PubNubError(.invalidArguments, additional: ["Invalid pushType parameter"])); return
+    }
+    pubnub.addPushChannelRegistrations(channels, for: deviceId, of: pushService) {
+      switch $0 {
+      case .success(let channels):
+        onSuccess(channels)
+      case .failure(let error):
+        onFailure(error)
+      }
+    }
+  }
+
+  func listPushChannels(
+    deviceId: Data,
+    pushType: String,
+    onSuccess: @escaping (([String]) -> Void),
+    onFailure: @escaping ((Error) -> Void)
+  ) {
+    guard let pushService = pushService(from: pushType) else {
+      onFailure(PubNubError(.invalidArguments, additional: ["Invalid pushType parameter"])); return
+    }
+    pubnub.listPushChannelRegistrations(for: deviceId, of: pushService) {
+      switch $0 {
+      case .success(let channels):
+        onSuccess(channels)
+      case .failure(let error):
+        onFailure(error)
+      }
+    }
+  }
+
+  func removeChannelsFromPush(
+    channels: [String],
+    deviceId: Data,
+    pushType: String,
+    onSuccess: @escaping (([String]) -> Void),
+    onFailure: @escaping ((Error) -> Void)
+  ) {
+    guard let pushService = pushService(from: pushType) else {
+      onFailure(PubNubError(.invalidArguments, additional: ["Invalid pushType parameter"])); return
+    }
+    pubnub.removePushChannelRegistrations(channels, for: deviceId, of: pushService) {
+      switch $0 {
+      case .success(let channels):
+        onSuccess(channels)
+      case .failure(let error):
+        onFailure(error)
+      }
+    }
+  }
+
+  func removeAllChannelsFromPush(
+    pushType: String,
+    deviceId: Data,
+    onSuccess: @escaping (() -> Void),
+    onFailure: @escaping ((Error) -> Void)
+  ) {
+    guard let pushService = pushService(from: pushType) else {
+      onFailure(PubNubError(.invalidArguments, additional: ["Invalid pushType parameter"])); return
+    }
+    pubnub.removeAllPushChannelRegistrations(for: deviceId, of: pushService) {
+      switch $0 {
+      case .success:
+        onSuccess()
+      case .failure(let error):
+        onFailure(error)
+      }
+    }
+  }
+}
