+++ Sources/PubNub/KMM/PubNubObjC+Subscribe.swift
@@ -1,0 +1,57 @@
+//
+//  PubNubObjC+Subscribe.swift
+//
+//  Copyright (c) PubNub Inc.
+//  All rights reserved.
+//
+//  This source code is licensed under the license found in the
+//  LICENSE file in the root directory of this source tree.
+//
+
+import Foundation
+
+// MARK: - Subscribed channels & channel groups
+
+@objc
+public extension PubNubObjC {
+  var subscribedChannels: [String] {
+    pubnub.subscribedChannels
+  }
+
+  var subscribedChannelGroups: [String] {
+    pubnub.subscribedChannelGroups
+  }
+}
+
+// MARK: - Subscribe & Unsubscribe
+
+@objc
+public extension PubNubObjC {
+  func subscribe(
+    channels: [String],
+    channelGroups: [String],
+    withPresence: Bool,
+    timetoken: Timetoken
+  ) {
+    pubnub.subscribe(
+      to: channels,
+      and: channelGroups,
+      at: timetoken,
+      withPresence: withPresence
+    )
+  }
+
+  func unsubscribe(
+    from channels: [String],
+    channelGroups: [String]
+  ) {
+    pubnub.unsubscribe(
+      from: channels,
+      and: channelGroups
+    )
+  }
+
+  func unsubscribeAll() {
+    pubnub.unsubscribeAll()
+  }
+}
