+++ Sources/PubNub/KMM/PubNubObjC+Time.swift
@@ -1,0 +1,30 @@
+//
+//  PubNubObjC+Time.swift
+//
+//  Copyright (c) PubNub Inc.
+//  All rights reserved.
+//
+//  This source code is licensed under the license found in the
+//  LICENSE file in the root directory of this source tree.
+//
+
+import Foundation
+
+// MARK: - Time
+
+@objc
+public extension PubNubObjC {
+  func time(
+    onSuccess: @escaping ((Timetoken) -> Void),
+    onFailure: @escaping ((Error) -> Void)
+  ) {
+    pubnub.time {
+      switch $0 {
+      case .success(let timetoken):
+        onSuccess(timetoken)
+      case .failure(let error):
+        onFailure(error)
+      }
+    }
+  }
+}
